requirements.txt
defusedxml==0.4.1
Django==1.9.7
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.3.3
google-api-python-client==1.5.1
httplib2==0.9.2
Markdown==2.6.6
oauth2client==2.2.0
oauthlib==1.1.2
Pillow==3.2.0
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyJWT==1.4.0
python-dateutil==2.5.3
python-openid==2.2.5
python-social-auth==0.2.19
python3-openid==3.0.10
pytz==2016.4
reportlab==3.3.0
requests==2.10.0
requests-oauthlib==0.6.1
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
urllib3==1.16